Update README.md
Hi team, because Image Viewer component now has become a community-supported component, we will remove its doc from the Docs site, but the content of doc should be moved to the Documentation tab on the component page so that users who need the doc can still access it. So I update this README file with the content of the doc. If you have any feedback, please feel free to contact Luyao Zhang from the Docs team.